component CalculatorPage creates and styles updated from diferents components
